requirements.txt
google-genai>=1.70.0
python-dotenv>=1.0.0
chromadb>=1.5.0
fastembed>=0.4.0
langchain-google-genai>=4.0.0
langchain-core>=1.0.0
langchain>=1.0.0
